Add ability to run assessments for file (#7)
* feat(PI-4625): add ability to run assessments on files

* feat(PI-4625): edit oapi spec for taskId numeric type bug

* avoid os.exit outside of command functions

* feat: add zerolog configurable logging

* logger should use stderr for warnings and erros, stdout otherwise

* config file path defaulted in cobra flag

* submitFile -> uploadFile

* specifically print the 2xx task status

* fix context passing for package

+++ internal/platformapi/client.oapi.go
@@ -8369,7 +8369,8 @@ type PostBuild2XX1 struct {
 	} `json:"status"`
 
 	// Task An integer identifying the underlying analysis tasks associated with this assessment.
-	Task float32 `json:"task"`
+	// TODO: Comment on upstream issue on float64 defaults
+	Task float64 `json:"task"`
 
 	// TaskErrorCode The error code describing the failure if an assessment failed.
 	TaskErrorCode *string `json:"task_error_code"`
